Decouple ASV tests from installed package
Move benchmark-specific tests under the ASV tree so wheel test
discovery no longer imports repository-only modules. Separate ASV track
adapters from metric collection and keep the shipped run_benchmark
contract test in newton.tests.
